Begin creating the game.
Create pickup classes for the player (they are functional but may need tweaking as we begin implementing). Create a player class and add various members and methods needed to by the pickups (should be using to get the player working with these methods). Add placeholder visual assets. Add a game state object and a fail condition.