Finish adding obstacles to the game.
Toy with drawing to give better visual depth to the game. Add in enemy animations and other artwork. Fix some of the collision boxes. Fix the scaling of some sprites. Limit the number of times an enemy will be moved to prevent overlap when spawning, preventing the game from entering an endless loop.